+++ GO/core/src/com/GO/core/manager/GoEngManager.java
@@ -13,11 +13,11 @@ public class GoEngManager {
     public static final byte CvP = 2; // cpu vs player
     public static final byte ONL = 3; // online
 
-    public static GoEngine startEngine(int mode, int nCells){
+    public static GoEngine startEngine(byte mode, byte nCells){
         switch (mode){
             case PvP: return new GoEngine(mode, nCells);
-            case PvC: return new GoBot((byte)mode, nCells);
-            case CvP: return new GoBot((byte)mode, nCells);
+            case PvC: return new GoBot(mode, nCells);
+            case CvP: return new GoBot(mode, nCells);
             case ONL: return new GoMultiplayer(mode, nCells);
             default: return null;
         }
